pyproject.toml
[project]
name = "tfm"
version = "0.1.0"
description = "Development of a System for Attacks Detection and Prevention in Web Applications"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "elasticsearch<9.0.0",
    "pytest>=9.0.2",
    "python-nmap>=0.7.1",
    "python-owasp-zap-v2-4>=0.1.0",
    "requests>=2.32.5",
]

[tool.uv]
package = true

[project.scripts]
tool = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
